# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Openclaw Agent OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type AgentStatus = (typeof AgentStatus)[keyof typeof AgentStatus];

export const AgentStatus = {
  idle: "idle",
  thinking: "thinking",
  searching: "searching",
  writing: "writing",
  delegating: "delegating",
  executing: "executing",
} as const;

export interface Agent {
  id: number;
  name: string;
  persona: string;
  toolsEnabled: string[];
  status: AgentStatus;
  lastActiveAt?: string | null;
  createdAt: string;
}

export interface CreateAgentBody {
  name: string;
  persona: string;
  toolsEnabled?: string[];
}

export interface UpdateAgentBody {
  name?: string;
  persona?: string;
  toolsEnabled?: string[];
}

export type UpdateAgentStatusBodyStatus =
  (typeof UpdateAgentStatusBodyStatus)[keyof typeof UpdateAgentStatusBodyStatus];

export const UpdateAgentStatusBodyStatus = {
  idle: "idle",
  thinking: "thinking",
  searching: "searching",
  writing: "writing",
  delegating: "delegating",
  executing: "executing",
} as const;

export interface UpdateAgentStatusBody {
  status: UpdateAgentStatusBodyStatus;
}

export interface Conversation {
  id: number;
  agentId: number;
  title: string;
  createdAt: string;
}

export type ConversationMessageRole =
  (typeof ConversationMessageRole)[keyof typeof ConversationMessageRole];

export const ConversationMessageRole = {
  user: "user",
  assistant: "assistant",
  tool: "tool",
} as const;

export interface SearchSource {
  title: string;
  url: string;
  snippet: string;
  favicon?: string | null;
}

export interface ConversationMessage {
  id: number;
  conversationId: number;
  role: ConversationMessageRole;
  content: string;
  sourcesJson?: SearchSource[] | null;
  timestamp: string;
}

export interface ConversationWithMessages {
  id: number;
  agentId: number;
  title: string;
  createdAt: string;
  messages: ConversationMessage[];
}

export interface AgentChatBody {
  content: string;
  conversationId?: number | null;
}

export interface AgentMessage {
  id: number;
  fromAgentId: number;
  toAgentId: number;
  threadId?: string | null;
  content: string;
  response?: string | null;
  timestamp: string;
}

export interface SendAgentMessageBody {
  toAgentId: number;
  content: string;
  threadId?: string | null;
}

export interface ActivityEntry {
  id: number;
  agentId?: number | null;
  agentName?: string | null;
  actionType: string;
  detail: string;
  timestamp: string;
}

export type VpsConfigAuthType =
  (typeof VpsConfigAuthType)[keyof typeof VpsConfigAuthType];

export const VpsConfigAuthType = {
  password: "password",
  key: "key",
} as const;

export interface VpsConfig {
  id: number;
  label: string;
  host: string;
  port: number;
  username: string;
  authType: VpsConfigAuthType;
  hasCredentials: boolean;
  createdAt: string;
}

export type SaveVpsConfigBodyAuthType =
  (typeof SaveVpsConfigBodyAuthType)[keyof typeof SaveVpsConfigBodyAuthType];

export const SaveVpsConfigBodyAuthType = {
  password: "password",
  key: "key",
} as const;

export interface SaveVpsConfigBody {
  label: string;
  host: string;
  port: number;
  username: string;
  authType: SaveVpsConfigBodyAuthType;
  password?: string | null;
  privateKey?: string | null;
}

export interface VpsTestResult {
  success: boolean;
  message: string;
}

export interface VpsStats {
  cpuPercent: number;
  memUsedMb: number;
  memTotalMb: number;
  diskUsedGb: number;
  diskTotalGb: number;
  uptime: string;
  loadAvg: string;
}

export interface VpsProcess {
  pid: number;
  user: string;
  cpu: string;
  mem: string;
  command: string;
}

export type VpsServiceType =
  (typeof VpsServiceType)[keyof typeof VpsServiceType];

export const VpsServiceType = {
  systemd: "systemd",
  pm2: "pm2",
  unknown: "unknown",
} as const;

export interface VpsService {
  name: string;
  status: string;
  type: VpsServiceType;
}

export type VpsFileEntryType =
  (typeof VpsFileEntryType)[keyof typeof VpsFileEntryType];

export const VpsFileEntryType = {
  file: "file",
  directory: "directory",
  symlink: "symlink",
} as const;

export interface VpsFileEntry {
  name: string;
  path: string;
  type: VpsFileEntryType;
  size: number;
  modifiedAt: string;
}

export interface VpsFileContent {
  path: string;
  content: string;
  encoding: string;
}

export interface WriteVpsFileBody {
  path: string;
  content: string;
}

export interface ExecVpsCommandBody {
  command: string;
  timeout?: number | null;
}

export interface VpsCommandResult {
  success: boolean;
  stdout: string;
  stderr: string;
  exitCode?: number | null;
}

export type AppSettingsSearchProvider =
  (typeof AppSettingsSearchProvider)[keyof typeof AppSettingsSearchProvider];

export const AppSettingsSearchProvider = {
  duckduckgo: "duckduckgo",
  brave: "brave",
} as const;

export interface AppSettings {
  aiModel: string;
  smtpHost?: string | null;
  smtpPort?: number | null;
  smtpUser?: string | null;
  smtpConfigured: boolean;
  webhookUrl?: string | null;
  searchProvider: AppSettingsSearchProvider;
  braveApiKeyConfigured: boolean;
}

export type SaveSettingsBodySearchProvider =
  (typeof SaveSettingsBodySearchProvider)[keyof typeof SaveSettingsBodySearchProvider];

export const SaveSettingsBodySearchProvider = {
  duckduckgo: "duckduckgo",
  brave: "brave",
} as const;

export interface SaveSettingsBody {
  aiModel?: string;
  smtpHost?: string | null;
  smtpPort?: number | null;
  smtpUser?: string | null;
  smtpPassword?: string | null;
  webhookUrl?: string | null;
  searchProvider?: SaveSettingsBodySearchProvider;
  braveApiKey?: string | null;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "512x512": "512x512",
  "256x256": "256x256",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

export type ListActivityParams = {
  limit?: number;
  agentId?: number;
};

export type ListVpsFilesParams = {
  path: string;
};

export type ReadVpsFileParams = {
  path: string;
};

export type DeleteVpsFileParams = {
  path: string;
};

export type TailVpsLogParams = {
  path: string;
  lines?: number;
};