# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Openclaw Agent OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create a new agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentBody"
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get an agent by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentBody"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /agents/{id}/status:
    patch:
      operationId: updateAgentStatus
      tags: [agents]
      summary: Update agent status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentStatusBody"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"

  # ── Agent Conversations ──────────────────────────────────────────────────────
  /agents/{id}/conversations:
    get:
      operationId: listAgentConversations
      tags: [conversations]
      summary: List conversations for an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
  /conversations/{id}:
    get:
      operationId: getConversation
      tags: [conversations]
      summary: Get a conversation with its messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteConversation
      tags: [conversations]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /agents/{id}/chat:
    post:
      operationId: agentChat
      tags: [conversations]
      summary: Send a message to an agent and receive a Perplexity-style streaming response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentChatBody"
      responses:
        "200":
          description: SSE stream of reasoning steps, sources, answer chunks, and follow-ups
          content:
            text/event-stream: {}
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Agent-to-Agent Messages ──────────────────────────────────────────────────
  /agents/{id}/messages:
    get:
      operationId: listAgentMessages
      tags: [agentMessages]
      summary: List messages sent to/from an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of agent messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentMessage"
    post:
      operationId: sendAgentMessage
      tags: [agentMessages]
      summary: Send a message from one agent to another
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAgentMessageBody"
      responses:
        "201":
          description: Message sent and delegation triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentMessage"

  # ── Activity Log ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Agent schemas ──────────────────────────────────────────────────────────
    Agent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        persona:
          type: string
        toolsEnabled:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [idle, thinking, searching, writing, delegating, executing]
        lastActiveAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - persona
        - toolsEnabled
        - status
        - createdAt
    CreateAgentBody:
      type: object
      properties:
        name:
          type: string
        persona:
          type: string
        toolsEnabled:
          type: array
          items:
            type: string
      required:
        - name
        - persona
    UpdateAgentBody:
      type: object
      properties:
        name:
          type: string
        persona:
          type: string
        toolsEnabled:
          type: array
          items:
            type: string
    UpdateAgentStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [idle, thinking, searching, writing, delegating, executing]
      required:
        - status

    # ── Conversation schemas ───────────────────────────────────────────────────
    Conversation:
      type: object
      properties:
        id:
          type: integer
        agentId:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - agentId
        - title
        - createdAt
    ConversationMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant, tool]
        content:
          type: string
        sourcesJson:
          nullable: true
          type: array
          items:
            $ref: "#/components/schemas/SearchSource"
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - timestamp
    ConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        agentId:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ConversationMessage"
      required:
        - id
        - agentId
        - title
        - createdAt
        - messages
    SearchSource:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        snippet:
          type: string
        favicon:
          type: string
          nullable: true
      required:
        - title
        - url
        - snippet
    AgentChatBody:
      type: object
      properties:
        content:
          type: string
        conversationId:
          type: integer
          nullable: true
      required:
        - content

    # ── Agent message schemas ──────────────────────────────────────────────────
    AgentMessage:
      type: object
      properties:
        id:
          type: integer
        fromAgentId:
          type: integer
        toAgentId:
          type: integer
        threadId:
          type: string
          nullable: true
        content:
          type: string
        response:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - fromAgentId
        - toAgentId
        - content
        - timestamp
    SendAgentMessageBody:
      type: object
      properties:
        toAgentId:
          type: integer
        content:
          type: string
        threadId:
          type: string
          nullable: true
      required:
        - toAgentId
        - content

    # ── Activity log schemas ───────────────────────────────────────────────────
...