# lib/api-zod/src/generated/types/createAgentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Openclaw Agent OS API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAgentBody {
  name: string;
  persona: string;
  toolsEnabled?: string[];
}