# lib/api-zod/src/generated/types/searchSource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Openclaw Agent OS API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchSource {
  title: string;
  url: string;
  snippet: string;
  favicon?: string | null;
}